[KOGITO-5723] Moving workflow utils to a different project
