+++ jbpm/process-workitems/src/main/java/org/kie/kogito/process/workitems/impl/expr/ParsedExpression.java
@@ -15,11 +15,9 @@
  */
 package org.kie.kogito.process.workitems.impl.expr;
 
-import java.util.Optional;
-
 public interface ParsedExpression {
 
     <T> T eval(Object context, Class<T> returnClass);
 
-    Optional<String> varName();
+    void assign(Object context, Object value);
 }
