+++ kogito-serverless-workflow/kogito-jsonpath-expression/pom.xml
@@ -23,6 +23,10 @@
       <artifactId>json-path</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>kogito-serverless-workflow-utils</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-api</artifactId>
       <scope>test</scope>
