+++ kogito-serverless-workflow/kogito-serverless-workflow-utils/src/main/java/org/kie/kogito/serverless/workflow/utils/WorkflowAppContext.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.serverless.workflow.parser.util;
+package org.kie.kogito.serverless.workflow.utils;
 
 import java.io.InputStream;
 import java.util.Properties;
