+++ kogito-serverless-workflow/pom.xml
@@ -17,6 +17,7 @@
   <modules>
     <module>kogito-serverless-workflow-builder</module>
     <module>kogito-serverless-workflow-runtime</module>
+    <module>kogito-serverless-workflow-utils</module>
     <module>kogito-jsonpath-expression</module>
     <module>kogito-jq-expression</module>
   </modules>
